+++ src/main/java/net/echo/brain4j/training/optimizers/impl/AdamW.java
@@ -41,7 +41,7 @@ public class AdamW extends Adam {
 
         for (Layer<?, ?> layer : layers) {
             for (Synapse synapse : layer.getSynapses()) {
-                double change = update(cacheHolder, synapse);
+                float change = (float) update(cacheHolder, synapse);
                 updater.acknowledgeChange(synapse, change);
             }
         }
