+++ src/main/java/net/echo/brain4j/training/optimizers/impl/GradientDescent.java
@@ -23,7 +23,7 @@ public class GradientDescent extends Optimizer {
     public void postIteration(StatesCache cacheHolder, Updater updater, List<Layer<?, ?>> layers) {
         for (Layer<?, ?> layer : layers) {
             for (Synapse synapse : layer.getSynapses()) {
-                double change = update(cacheHolder, synapse);
+                float change = (float) update(cacheHolder, synapse);
                 updater.acknowledgeChange(synapse, change);
             }
         }
